Rewrite zshrc and tmux.conf
Remove oh-my-zsh and put in a custom pretty prompt.  I will add more to
my zshrc as I find functionality from oh-my-zsh I need.  I wasn't using
many of the plugins from oh-my-zsh and the ones that I were I wasn't
looking at.

The tmux.conf is super basic at the moment.  As with the zshrc I'll add
in customization as a go.  There is a bit of an issue with true color in
tmux sessions.  If I turn it on it forces tig and other applications which
may not support true color to use true color, and if it's off vim looks
weird.  I'll have to do more with this.

Also.  I need to add back my per file customization to vimrc